backend/pyproject.toml
[tool.poetry]
name = "praychain-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
openai-whisper = "^20240930"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.0"
motor = "^3.6.0"
transformers = "^4.47.1"
torch = "^2.6.0"
requests = "^2.32.3"
python-multipart = "^0.0.6"
protobuf = ">=4.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
